(문준식/front/comment) 댓글 기능 오류 수정

+++ assets/js/review.js
@@ -39,7 +39,7 @@ const getLogInUserId = async () => {
 }
 
 const currentMemberId = await getLogInUserId();
-const currentMemberRole = getLogInUserRole();
+const currentMemberRole = await getLogInUserRole();
 
 
 
@@ -109,12 +109,12 @@ const speadReviews = async (reviews, currentMemberId, role) => {
 
     // 수정 버튼 클릭 시 수정 입력란을 보여주는 이벤트 추가
     commentElement.find('.edit-comment-btn').on('click', function () {
-      editComment(commentElement, reviewId);
+      editComment(commentElement, reviewId, currentMemberRole);
     });
 
     // 삭제 버튼 클릭 시 댓글 삭제하는 이벤트 추가
     commentElement.find('.delete-comment-btn').on('click', function () {
-      deleteComment(reviewId, commentElement);
+      deleteComment(reviewId, commentElement, currentMemberRole);
     });
 
     reviewDiv.append(commentElement);
@@ -135,11 +135,12 @@ const createReview = async (petSitterId, role) => {
 
     if (!token) {
       alert('로그인 후 이용 가능합니다.');
-      window.location.href = "/user-sign-in";
+      return window.location.href = "/user-sign-in";
     };
 
     if (role !== 'user') {
-      alert('리뷰는 유저만 남길 수 있습니다.')
+      console.log('role: ', role);
+      return alert('리뷰는 유저만 남길 수 있습니다.')
     }
 
     const inputComment = $('.comment-input').val();
@@ -172,12 +173,17 @@ const createReview = async (petSitterId, role) => {
 /* 버튼 클릭 시 리뷰 생성하기 */
 $('.comment-btn').on('click', async (event) => {
   event.preventDefault()
-  createReview(petSitterId)
+  createReview(petSitterId, currentMemberRole)
 });
 
 
 /* 리뷰 수정하기 */
-export const editComment = async (commentElement, reviewId) => {
+export const editComment = async (commentElement, reviewId, currentMemberRole) => {
+  if (currentMemberRole !== 'user') {
+    alert("리뷰 수정은 작성한 유저 본인만 가능합니다.")
+    return;
+  }
+
   const commentText = commentElement.find('.comment-text').text();
 
   // 기존의 수정, 삭제 버튼 숨기기
@@ -257,7 +263,12 @@ export const editComment = async (commentElement, reviewId) => {
 
 
 // 댓글 삭제하는 함수
-export const deleteComment = async function (reviewId, commentElement) {
+export const deleteComment = async function (reviewId, commentElement, currentMemberRole) {
+  if (currentMemberRole !== 'user') {
+    alert("댓글 삭제는 작성한 유저 본인만 가능합니다.")
+    return;
+  }
+
   try {
     await fetch(`/api/pet-sitters/reviews/${reviewId}`, {
       method: 'DELETE',
